some import cleanups

+++ Meowijuana/Pathfinding.cpp
@@ -1,19 +1,12 @@
 ﻿#include "pch.h"
-#include "Entity.hpp"
-
+#include "Pathfinding.hpp"
 
 namespace Path {
-
 	Vect randPoint() {
-
-		float direction = (AERandFloat() * 360);
-		float distance = (AERandFloat() * 200);
+		float direction = static_cast<float>(AERandFloat() * 360.0f);
+		float distance = static_cast<float>(AERandFloat() * 200.0f);
 
 		// move in a random direction for a random distance
 		return Vect{ distance * AECos(direction), distance * AESin(direction) };
 	}
-
-
-
 }
-
